Fix wrong object type after create a transaction (#252)
* customer: fix customer type on API response

When the SDK create the transaction object from the API response the
customer object is showing like a stdClass. This commit fix this by
switching this stdClass to the Customer object.

* card: fix card type on API response

When the SDK create the transaction object from the API response the
card object is showing like a stdClass. This commit fix this by
switching this stdClass to the Card object.

* tests: fix tests to not access id property directly

* tests: fix tests failing

The bulk anticipation tests fails in some days of the week because it
count +5 days and +6 days from current date time, and the anticipation
must be in a business day. This PR fix this by validating the week day
and changing the count if the day is a weekend.