feature: Client + Http (#276)
* client: add client

* client: add apiKey validation

* codesniffer: fix codesniffer errors

* composer: remove Sdk from autoload namespace

* endpoints: remove transaction endpoint

* client: add request

This commit adds guzzle and the request functionality to the client, to
be possible to retrive information from the pagar.me API.

* environment: define a standard for Composer image.

* response: success && failure

* client: consuming resources from ResponseHandler.

* exception: new class to define an error layer between Client and final
user.

* ci: fixes to PSR-2 verification

Previously at #277, we fixed both Travis and CircleCI settings. With
the changes presented here, we're introducing the inspection of the
`tests/` directory, as we have test files to inspect :wink:

* review: fix URI usage for query parameters.

* review: required fixes

* response-handler: change error handling

Today only one exception type are throwed by the ResponseHandler, this
commit change the exception handling by adding different exception type
based on whether the request failed or the API failed.