feature/query-string-on-gets (#293)
* client: API key usage for all requests

Every request to Pagar.me's API requires the API Key. So, in order to
follow this rule, we're assigning this value as a param on the query
string; but, instead of add the value as a suffix for all URI's, we'll
use the Guzzle query resource.

* endpoints: allow search on endpoints using query params

* endpoints: docblock fixes

These docblocks were broken, so a small fix was required :wink:

* tests: refactor

As we're shipping the query string communication for each endpoint, we
need to add tests that cover scenarios where the user may want to use
query params and where it may won't.

Plus, we're fixing some details in our tests.