Adiciona header User-Agent aos requests feitos pelo SDK (#309)
* client: build user-agent string

Implement an user-agent string builder to be user during requests to
Pagar.me's API. It'll be usefull to indentify applications requests.

Co-authored-by: Murilo Souza <murilohns@users.noreply.github.com>

* client: inform user-agent on requests

Since a user-agent string is builded, it should be present on every
request. It add the builded string to the client.

Co-authored-by: Murilo Souza <murilohns@users.noreply.github.com>

* client: refactor the code

Create a constant for custom user-agent header. Also, refactor the test
to work well in different php versions.

Co-authored-by: Murilo Souza <murilohns@users.noreply.github.com>

* client: avoid php send an undefined index notice

Inform a blank user-agent if it's not informed from the client during
instantiation

* doc: add how to inform custom headers

Add how developers could inform custom headers on every request

* travis-ci: remove travis-ci config file

Remove travis since the project is using circle-ci

* refactor: remove irrelevant user_agent property

Since the user-agent info won't be used by developers after it sent to
sdk, there's no reason to maintain a property on Client class. It
removes the property and its getter method and remove tests that was
used to validate that.